Update the tests.
