--- test/AST-error/kernel_invalid_argument-1.cpp
@@ -1,32 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2022 - 2024 NVIDIA Corporation & Affiliates.                  *
- * All rights reserved.                                                        *
- *                                                                             *
- * This source code and the accompanying materials are made available under    *
- * the terms of the Apache License 2.0 which accompanies this distribution.    *
- ******************************************************************************/
-
-// RUN: cudaq-quake %cpp_std -verify %s
-
-#include <cudaq.h>
-#include <string>
-#include <tuple>
-
-// expected-error@+1{{kernel argument type not supported}}
-void prepQubit(std::string basis, cudaq::qubit &q) __qpu__ {}
-
-// expected-error@+1{{kernel argument type not supported}}
-void RzArcTan2(bool input, std::string basis) __qpu__ {
-  cudaq::qubit aux;
-  cudaq::qubit resource;
-  cudaq::qubit target;
-  if (input) {
-    x(target);
-  }
-  prepQubit(basis, target);
-}
-
-int main() {
-  RzArcTan2(true, {});
-  return 0;
-}
